client/requirements.txt
websockets==14.1
pyaudio==0.2.14
openwakeword==0.6.0
numpy<2
sounddevice==0.5.1
speechbrain==1.0.2
huggingface_hub>=0.20.0
torch==2.6.0
torchaudio==2.6.0
pydantic==2.10.4
python-dotenv==1.0.1
httpx==0.28.1
silero-vad>=5.1
onnxruntime>=1.17.0
faster-whisper>=1.0.0
